% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_con.R
\name{epi_con}
\alias{epi_con}
\title{Simulate epidemics}
\usage{
epi_con(birth_rate, death_rate, stop_time, x0 = 1)
}
\arguments{
\item{birth_rate}{birth rate of the epidemic.}

\item{death_rate}{death rate of the epidemic.}

\item{stop_time}{number of time points to run the epidemic simulation for.}

\item{x0}{initial number of infectives.}
}
\value{
data frame with: Index, Parent Index, Infection Status, Infection Time, Removal Time
}
\description{
Simulates a birth-death epidemic with constant birth and death rates and x0 infected on day 0.
}
\examples{
epi_con(birth_rate = 0.2, death_rate = 0.1, stop_time = 50)
}
