% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_fn.R
\name{epi_fn}
\alias{epi_fn}
\title{Simulate epidemics with birth rate varying over time.}
\usage{
epi_fn(birth_rate_fn, death_rate, stop_time, x0 = 1)
}
\arguments{
\item{birth_rate_fn}{function of the birth rate over time of the epidemic.}

\item{death_rate}{death rate of the epidemic.}

\item{stop_time}{number of time points to run the epidemic simulation for.}

\item{x0}{initial number of infectives.}
}
\value{
data frame with: Index, Parent Index, Infection Status, Infection Time, Removal Time
}
\description{
Simulates a birth-death epidemic with varying birth rate and constant death rates and x0 infected on day 0.
}
\examples{
epi_fn(birth_rate_fn = linear, death_rate = 0.1, stop_time = 50)
}
