% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_fns.R
\name{epi_fns}
\alias{epi_fns}
\alias{constant}
\alias{change_point}
\alias{peak}
\alias{increase}
\alias{decrease}
\title{A collection of functions for the birth rate in used to generate epidemics}
\usage{
constant(t)

change_point(t)

peak(t)

increase(t)

decrease(t)
}
\arguments{
\item{t}{time}
}
\value{
function
}
\description{
Epidemic functions
}
