% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_data.R
\name{genetic_data}
\alias{genetic_data}
\title{Genetic Data}
\usage{
genetic_data(ptree, stop_time, ptree_lag = 0)
}
\arguments{
\item{ptree}{object of class phylo.}

\item{stop_time}{number of times the epidemic simulation was run for.}

\item{ptree_lag}{number of time points in the past the most recent leaf was sampled.}
}
\value{
data frame with 3 columns and N+1 rows: Column 1 is time, column 2 is the number of lineages at the end of that time, column 3 is the number of coalescences on that time.
}
\description{
Discretises a phylogenetic tree.
}
\examples{
genetic_data(ptree = sample_tree, stop_time = 50, ptree_lag = 3)
}
