% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_phylo.R
\name{sample_phylo}
\alias{sample_phylo}
\title{Sample Phylogenetic Tree}
\usage{
sample_phylo(ptree, ptree_lag = 0, pi0, pi1)
}
\arguments{
\item{ptree}{object of class phylo.}

\item{pi0}{probability of sampling a leaf at the present time.}

\item{pi1}{probability of sampling a leaf before the present time.}

\item{time}{how many time points in the past the most recent leaf is.}
}
\value{
object of class phylo; the tree keeping only the sampled leaves.
}
\description{
Samples and keeps leaves with probability pi.
}
\examples{
sample_phylo(ptree = full_tree, pi0 = 0, pi1 = 0.01)
}
