% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{sir}
\alias{sir}
\title{Sample Importance Resample - Adaptive}
\usage{
sir(
  n_particles,
  ess_threshold = n_particles/2,
  x0 = 1,
  death_rate,
  sigma,
  reporting_prob,
  sample_prevalence,
  genetic_data,
  resampling_scheme = "systematic",
  backward_sim = TRUE
)
}
\arguments{
\item{n_particles}{number of particles used in the sampling.}

\item{ess_threshold}{threshold of ESS below which triggers resampling.}

\item{x0}{prevalence at time 0.}

\item{death_rate}{death rate of the epidemic.}

\item{sigma}{standard deviation of the linear gaussian proposal for birth rates.}

\item{reporting_prob}{reporting probability of cases observed.}

\item{sample_prevalence}{data frame of observed prevalence per day.}

\item{genetic_data}{data frame of time, number of lineages and number of coalescences.}

\item{resampling_scheme}{"multinomial" or "systematic".}

\item{backward_sim}{logical; if TRUE, uses backward simulation.}
}
\value{
log-likelihood
}
\description{
Implements an adaptive SIR algorithm and returns an approximated log-likelihood. Prevalence proposals are negative binomial and birth rate proposals are linear Gaussian.
}
\examples{
sir(n_particles = 100, x0 = 10, death_rate = 0.1, sigma = 0.1, reporting_prob = 0.1, sample_prevalence = noisy_prev, genetic_data = gen_data)
}
